% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFlights.R
\name{fetchFlights}
\alias{fetchFlights}
\title{Return all the departing flights from a city during a interval of time}
\usage{
fetchFlights(city, date, interval = 1, ryanairCities)
}
\arguments{
\item{city}{the deprating city}

\item{date}{the date of departure}

\item{interval}{the number of days to be considered from \code{date}. Default to 1
means that only one day is considered}

\item{ryanairCities}{the data.frame of all the served cities}
}
\description{
Return all the departing flights from a city during a interval of time
}
