% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRoute.R
\name{findRoute}
\alias{findRoute}
\title{Solve the problem of finding a route between two cities of the Ryanair dataset}
\usage{
findRoute(from, to, date, interval = 1, max.stops = 5)
}
\arguments{
\item{from}{the departing city}

\item{to}{the arrival city}

\item{date}{the departing day}

\item{interval}{how many days to consider from date. Default 1 means that departure
is exactly at \code{date}}
}
\description{
Solve the problem of finding a route between two cities of the Ryanair dataset
}
