% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRyanairCities.R
\name{getRyanairCities}
\alias{getRyanairCities}
\title{Get all the cities served by Ryanair}
\usage{
getRyanairCities(url = "https://desktopapps.ryanair.com/en-gb/res/stations")
}
\arguments{
\item{url}{the url of the Ryanair api}
}
\value{
A data.frame with all the cities
}
\description{
Get all the cities served by Ryanair
}
